visual moved to visual.py module, text related to utils.py
