+++ scr/evaluate.py
@@ -18,12 +18,12 @@ from sklearn.metrics import mean_absolute_error
 import os
 from utils import load_env
 from dataset import construct_test_set
-import re
-import matplotlib.pyplot as plt
 from PIL import Image
 import torchvision.transforms as transforms
 import numpy as np
 from torchvision.transforms import ToPILImage
+from visual import visualize_evaluation_results
+from utils import get_loss_key, read_log_results
 to_pil_image = ToPILImage()
 
 LOSS = 'FocalTverskyLoss' # change the results log ['DiceLoss', 'TverskyLoss', 'FocalTverskyLoss', 'BCEWithLogitsLoss']
@@ -135,16 +135,6 @@ test_epoch = smp.utils.train.ValidEpoch(
     device=DEVICE,
 )
 
-
-def camel_to_snake(name):
-    name = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
-    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', name).lower()
-
-
-def get_loss_key(loss_name):
-    key_base = camel_to_snake(loss_name[:-4])  # Remove 'Loss' and convert
-    return key_base + "_loss_weighted"  # Append '_loss_weighted' as in train.py logs
-
 # Convert the image to a pytorch tensor
 def image_tensor(img):
     """
@@ -180,111 +170,6 @@ def image_tensor(img):
         return image
     else:
         raise TypeError("Input must be np.ndarray or PIL.Image")
-
-
-def mask_to_rgba(mask, color="red", opacity=0.5):
-    MASK_COLORS = ["red", "green", "blue", "yellow", "magenta", "cyan"]
-    assert color in MASK_COLORS
-    assert mask.ndim == 3 or mask.ndim == 2
-
-    h = mask.shape[0]
-    w = mask.shape[1]
-    zeros = np.zeros((h, w))
-    ones = mask.reshape(h, w)
-    if color == "red":
-        return np.stack((ones, zeros, zeros, ones * opacity), axis=-1)
-    elif color == "green":
-        return np.stack((zeros, ones, zeros, ones * opacity), axis=-1)
-    elif color == "blue":
-        return np.stack((zeros, zeros, ones, ones * opacity), axis=-1)
-    elif color == "yellow":
-        return np.stack((ones, ones, zeros, ones * opacity), axis=-1)
-    elif color == "magenta":
-        return np.stack((ones, zeros, ones, ones * opacity), axis=-1)
-    elif color == "cyan":
-        return np.stack((zeros, ones, ones, ones * opacity), axis=-1)
-
-def read_log_results(logs_file_path):
-    with open(logs_file_path, 'r') as logs_file:
-        lines = logs_file.readlines()  # Read all lines into a list
-        return lines[-1]  # Return the last line
-       
-def visualize_evaluation_results(image_tensor, pred1, pred2, gt1, gt2, results_text, threshold=0.5):
-    """
-    Visualizes the segmentation results by the model with thresholding and color overlay, alongside ground truth masks.
-    
-    Parameters:
-    - image_tensor: A PyTorch tensor of the image to be visualized. 
-    - pred1, pred2: Model predictions for breast area and density.
-    - gt1, gt2: Ground truth masks for breast area and density.
-    - threshold: The threshold for converting probability maps to binary masks.
-    """
-    
-    # Apply threshold to convert predictions to binary masks
-    pred1_binary = (pred1 > threshold).float().squeeze().cpu().numpy()
-    pred2_binary = (pred2 > threshold).float().squeeze().cpu().numpy()
-
-    # Convert ground truths to RGBA for visualization
-    gt1_colored = mask_to_rgba(gt1.squeeze().cpu().numpy(), color="cyan")
-    gt2_colored = mask_to_rgba(gt2.squeeze().cpu().numpy(), color="magenta")
-
-    # Convert predictions to RGBA for visualization
-    pred1_colored = mask_to_rgba(pred1_binary, color="green")
-    pred2_colored = mask_to_rgba(pred2_binary, color="red")
-
-    # Convert the original image tensor to PIL for consistent visualization
-    input_pil = to_pil_image(image_tensor.cpu())
-
-    # Visualization
-    fig, axes = plt.subplots(2, 3, figsize=(16, 9))
-    
-    # Original Image
-    axes[0, 0].imshow(input_pil, cmap='gray')
-    axes[0, 0].set_title('Original Image')
-    axes[0, 0].axis('off')
-
-    # Prediction Breast Area
-    axes[0, 1].imshow(input_pil, cmap='gray')
-    axes[0, 1].imshow(pred1_colored, interpolation='none')
-    axes[0, 1].set_title('Prediction: Breast Area')
-    axes[0, 1].axis('off')
-
-    # Prediction Density
-    axes[0, 2].imshow(input_pil, cmap='gray')
-    axes[0, 2].imshow(pred2_colored, interpolation='none')
-    axes[0, 2].set_title('Prediction: Density')
-    axes[0, 2].axis('off')
-
-    # Ground Truth Breast Area
-    axes[1, 1].imshow(input_pil, cmap='gray')
-    axes[1, 1].imshow(gt1_colored, interpolation='none')
-    axes[1, 1].set_title('Ground Truth: Breast Area')
-    axes[1, 1].axis('off')
-
-    # Ground Truth Density
-    axes[1, 2].imshow(input_pil, cmap='gray')
-    axes[1, 2].imshow(gt2_colored, interpolation='none')
-    axes[1, 2].set_title('Ground Truth: Density')
-    axes[1, 2].axis('off')
-
-    # Format and display results text
-    axes[1, 0].clear()
-    axes[1, 0].axis('off')
-    results_lines = results_text.split('\t')  # Split the results text into separate lines
-
-    if results_lines:
-        results_lines = results_lines[1:]
-    
-    # Remove "dsc_plus_plus_" from the first three lines
-    for i in range(min(3, len(results_lines))):
-        results_lines[i] = results_lines[i].replace('dsc_plus_plus_', '')
-
-    formatted_text = '\n'.join(results_lines)
-    axes[1, 0].text(0.05, 0.75, formatted_text, ha='left', va='top', fontsize=12, wrap=True, transform=axes[1, 0].transAxes)
-
-    plt.show()
-
-
     
 # Define the maximum number of images to show
 max_images_to_show = 1  # You can adjust this number as needed
@@ -336,10 +221,10 @@ results_text = read_log_results(f"test_output/evaluation/{MODEL_NAME}.txt")
 # After evaluation and logging, visualize the stored images and their ground truths
 print("Visualizing results...")
 for img, gt_breast_area, gt_density in zip(images_to_visualize, ground_truths_breast_area, ground_truths_density):
-    # Assuming 'pred1' and 'pred2' are the predictions for the current 'img':
+
     # This part needs to call the prediction again for each 'img', or you adjust the loop to store predictions
     # The following is a placeholder for the actual prediction call
-    pred1, pred2 = model(img.unsqueeze(0).to(DEVICE))  # You need to ensure this matches how you handle your model predictions
+    pred1, pred2 = model(img.unsqueeze(0).to(DEVICE))
 
     # Convert predictions to binary masks based on a threshold, e.g., 0.5
     pred1_binary = (pred1 > 0.5).float()
